server
DTO  classses are created

+++ src/main/java/com/planner/journeyplanner/like/LikeRepository.java
@@ -8,7 +8,7 @@ public interface LikeRepository extends JpaRepository<Like, Long> {
 
     Long countByJourneyId(Long journeyId);
 
-    Optional<Like> findByUserIdAndJourneyId(Long userId, Long journeyId);
+    Optional<Like> findByUserIdAndJourneyId(Integer userId, Long journeyId);
 
     void delete(Like like);
 }
